Refactor authentication modules and exception handling
Several modifications were applied to clean up and optimize the code. The 'handlerExceptionResolver' parameter was removed from the JwtAuthenticationFilter constructor as it was not used. All CORS related code in 'SecurityConfiguration.java' was also removed because it was commented out and not in use. Simplified 'ValidationExceptionHandler' by removing unnecessary 'handleMethodValidationException' method and unused imports. Made changes to 'AbstractCheckedException' and 'CustomUserDetails' to handle serialization and object comparison properly.